+++ src/firecrest-streamer/src/streamer/streamer_server.py
@@ -1,13 +1,15 @@
 import asyncio
 import base64
-from enum import Enum
+import click
 import http
 import json
 import signal
 import websockets
+from enum import Enum
+from dataclasses import dataclass, replace
+from streamer.streamer_core import ConsoleReporter, stream_receive, stream_send
+from typing import List, Tuple
 from websockets.asyncio.server import serve
-import click
-from streamer.streamer_core import stream_send, stream_receive
 
 CHUNK_SIZE = 5 * 1024 * 1024  # 5 MiB
 
@@ -17,63 +19,71 @@ class Operation(Enum):
     receive = "receive"
 
 
-operation: Operation = None
-target: str = None
-secret: str = None
-port_range: tuple[int, int] = None
-ips: list[str] = None
-host: str = None
-wait_timeout: int = None
-inbound_transfer_limit: int = None
-timeout_handle: asyncio.Handle = None
+@dataclass
+class StreamConfig:
+    operation: Operation
+    target: str
+    secret: str
+    port_range: Tuple[int, int]
+    ips: List[str]
+    host: str
+    wait_timeout: int
+    inbound_transfer_limit: int
 
 
-async def server_receive(websocket: websockets.asyncio.server.ServerConnection):
-    global operation, target, inbound_transfer_limit
-    print("Client connected.")
-    try:
-        await stream_receive(websocket, target, inbound_transfer_limit)
-    except Exception as e:
-        print(f"An error occurred: {e}")
-    websocket.server.close()
+async def stream(config: StreamConfig):
+    timeout_handle: asyncio.Handle = None
+    reporter = ConsoleReporter()
 
+    def process_request(connection, request):
+        nonlocal timeout_handle
+        if "Authorization" not in request.headers:
+            return connection.respond(
+                http.HTTPStatus.UNAUTHORIZED, "Missing Authorization header\n"
+            )
 
-async def server_send(websocket):
-    global target
-    print("Client connected.")
-    try:
-        await stream_send(websocket, target)
-    except Exception as e:
-        print(f"An error occurred: {e}")
-    websocket.server.close()
+        authorization = request.headers["Authorization"]
+        if authorization is None:
+            return connection.respond(http.HTTPStatus.UNAUTHORIZED, "Missing token\n")
 
+        token = authorization.split("Bearer ")[-1]
+        if token is None or token != config.secret:
+            return connection.respond(http.HTTPStatus.FORBIDDEN, "Invalid secret\n")
 
-def process_request(connection, request):
-    global secret, timeout_handle
-    if "Authorization" not in request.headers:
-        return connection.respond(
-            http.HTTPStatus.UNAUTHORIZED, "Missing Authorization header\n"
-        )
+        if timeout_handle:
+            timeout_handle.cancel()
 
-    authorization = request.headers["Authorization"]
-    if authorization is None:
-        return connection.respond(http.HTTPStatus.UNAUTHORIZED, "Missing token\n")
+    async def server_receive(websocket: websockets.asyncio.server.ServerConnection):
+        reporter.info("Client connected.")
+        try:
+            await stream_receive(
+                websocket,
+                config.target,
+                config.inbound_transfer_limit,
+                reporter=reporter,
+            )
+        except Exception as e:
+            reporter.error(f"An error occurred: {e}")
+        websocket.server.close()
 
-    token = authorization.split("Bearer ")[-1]
-    if token is None or token != secret:
-        return connection.respond(http.HTTPStatus.FORBIDDEN, "Invalid secret\n")
+    async def server_send(websocket):
+        reporter.info("Client connected.")
+        try:
+            await stream_send(websocket, config.target, reporter=reporter)
+        except Exception as e:
+            reporter.error(f"An error occurred: {e}")
+        websocket.server.close()
 
-    timeout_handle.cancel()
-
-
-async def stream():
-    global secret, port_range, ips, host, wait_timeout, timeout_handle
-    start_port, end_port = port_range
+    start_port, end_port = config.port_range
     for port in range(start_port, end_port + 1):
         try:
             async with serve(
-                server_receive if operation == Operation.receive else server_send,
-                host,
+                (
+                    server_receive
+                    if config.operation == Operation.receive
+                    else server_send
+                ),
+                config.host,
                 port,
                 max_size=int(
                     CHUNK_SIZE * 1.25
@@ -82,25 +92,26 @@ async def stream():
                 ping_timeout=None,
                 process_request=process_request,
             ) as server:
-                print(f"Server is listening on ws://{host}:{port}")
+                reporter.info(f"Server is listening on ws://{config.host}:{port}")
                 coordinates = {
                     "ports": [start_port, end_port],
-                    "ips": ips,
-                    "secret": secret,
+                    "ips": config.ips,
+                    "secret": config.secret,
+                    "operation": config.operation.value,
                 }
                 encoded = base64.urlsafe_b64encode(
                     json.dumps(coordinates).encode("utf-8")
                 ).decode("utf-8")
 
-                print(f"Use these coordinates to connect: {encoded}", flush=True)
+                reporter.info(f"Use these coordinates to connect: {encoded}")
 
                 loop = asyncio.get_running_loop()
                 loop.add_signal_handler(signal.SIGTERM, server.close)
-                timeout_handle = loop.call_later(wait_timeout, server.close)
+                timeout_handle = loop.call_later(config.wait_timeout, server.close)
                 await server.wait_closed()
             break
         except OSError:
-            print(f"Server unable to bing on port: {port}")
+            reporter.error(f"Server unable to bing on port: {port}")
             continue
 
 
@@ -143,31 +154,36 @@ async def stream():
     help="Limit how much data can be received (in bytes)",
     default=5 * 1024 * 1024 * 1024,  # 5GB
 )
-def server(_secret, _ips, _host, _port_range, _wait_timeout, _inbound_transfer_limit):
-    global secret, port_range, ips, wait_timeout, inbound_transfer_limit, host
-    secret = _secret
-    port_range = _port_range
-    ips = _ips
-    host = _host
-    wait_timeout = _wait_timeout
-    inbound_transfer_limit = _inbound_transfer_limit
+@click.pass_context
+def server(
+    ctx, _secret, _ips, _host, _port_range, _wait_timeout, _inbound_transfer_limit
+):
+    ctx.ensure_object(dict)
+    ctx.obj = StreamConfig(
+        operation=None,
+        target=None,
+        secret=_secret,
+        port_range=_port_range,
+        ips=_ips,
+        host=_host,
+        wait_timeout=_wait_timeout,
+        inbound_transfer_limit=_inbound_transfer_limit,
+    )
 
 
 @server.command()
 @click.option("--path", help="The target path of the file to be sent.", required=True)
-def send(path):
-    global operation, target
-    operation = Operation.send
-    target = path
-    asyncio.run(stream())
+@click.pass_context
+def send(ctx, path):
+    config = replace(ctx.obj, operation=Operation.send, target=path)
+    asyncio.run(stream(config))
 
 
 @server.command()
 @click.option(
     "--path", help="The target path of the file to be received.", required=True
 )
-def receive(path):
-    global operation, target
-    operation = Operation.receive
-    target = path
-    asyncio.run(stream())
+@click.pass_context
+def receive(ctx, path):
+    config = replace(ctx.obj, operation=Operation.receive, target=path)
+    asyncio.run(stream(config))